requirements.txt
Django
requests
mysqlclient
djangorestframework